fix: handlePriceChange to ProfileCategories
