--- src/features/tutorApplication/lib/type.ts
@@ -1,15 +1,0 @@
-
-
-export interface ProfileCategoriesProps {
-  category: {
-    ageCategory: string;
-    price: string;
-  };
-  data: {
-    value: string;
-    label: string;
-  }[];
-  isLast?: boolean;
-  onAddCategory?: () => void;
-  className?: string;
-}
